ez_project/requirements.txt
Django>=4.2
djangorestframework>=3.14
python-decouple>=3.8
Pillow>=10.0.0
